requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-crispy-forms==1.6.0
django-filter==0.13.0
djangorestframework==3.3.3
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
